+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added image caption template system (#6, #154)  
 * Added Angular material to the demo page (#212)
 * Added configuration for caption colors (#213, #212)